don't remove newlines from WebDebugToolbar JavaScript
I had to debug an issue in the JavaScript of the WebDebugToolbar and it was very annoying that the code was
"pseudo-minified" just by removing the newlines.

I don't see any win in removing the newlines. The Toolbar is only used for development purposes and it doesn't even minify the JavaScript.