Add a check to make sure the session can hold the data that is being saved
We faced a problem in our app where the session data was being reset without any reason. 
We traced it down to the data being truncated when saving, so when it was read, being corrupt, it was reset.
There was no error anywhere, so we thought this check would be nice to have.