Support omitting the <target> node in an .xlf file.
According to the XLIFF 1.2 specifications, the <target> node can be omitted:

http://docs.oasis-open.org/xliff/v1.2/os/xliff-core.html#trans-unit

In the source language file, there is no need for a <target> node, because source and target are the same (provided that the translation key is stored in the "resname" attribute). The translator should fall back to the source string in this case, which this commit changes accordingly.