[validator] Disable Explicit Class Name Group
If "class name" group is **explicity** configured in a constraint, the
constraint will not be added to "Default" group.

The following example reproduces this issue:

```php
class Foo
{
    /**
     * @Assert\NotNull(
     *     groups = {"Foo"}
     * )
     */
    public $nameInFooGroup;

    /**
     * @Assert\NotNull
     */
    public $nameWithoutGroup;

    /**
     * @Assert\NotNull(
     *     groups = {"Default"}
     * )
     */
    public $nameInDefaultGroup;

}
```
Constraints in "Foo" Group:
1.  NotNull constraint on nameWithoutGroup property
2.  NotNull constraint on nameInDefaultGroup property
3.  NotNull constraint on nameInFooGroup property

Constraints in "Default" Group:
1.  NotNull constraint on nameWithoutGroup property
2.  NotNull constraint on nameInDefaultGroup property

The NotNull constraint on nameInFooGroup property is not added to "Default" group.
This is incorrect because "Default" group and "class name" group should be identical
if neither "group sequence" nor "embed object" is used.

The implementation of "class name" group didn't follow the JSR-303 spec.
For class X, group X can not be explicitly defined, because all explicit groups
are implemented as inner interfaces of class X and none of the interfaces can use X
as its name, because ``class X is already defined``.
This explains why, in the formal group definition, group X does not include constraints
explicitly assigned to group X, because that's impossible in java.

> The group X contains the following constraints:
   * all Default constraints hosted on X
   * all Default constraints hosted on interfaces of X
   * ...

This is the official documentation for "Default" group.
> Constraints in the Default group of a class are the constraints that have either no
explicit group configured or that are **configured to a group equal to the class name** or
the string Default.

It does not clarify the meaning of "**configured to a group equal to the class name**":
does it mean "implicitly configured" or "explicitly configured"?

If the class name group is not explicitly defined, then everything is fine. Unfortunately,
in Symfony, it's also possible to explicitly configure the "class name" group for constraints:

```php
...
class User
{
    /**
     * @Assert\NotNull(
     *      groups = {"User"}
     * )
    */
    public $name;
    ...
}
```
In this case, the "NotNull" constraint on name property won't be added to Default group
and that's the issue I am trying to fix.

Follow the JSR-303 spec to forbid the configuration of explicit "class name" group and
when an explicit class name group is found, a GroupDefinitionException exception shall be thrown.

| Q             | A
| ------------- | ---
| Bug fix?      | yes
| New feature?  | no
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | yes
| Fixed tickets | n/a
| License       | MIT
| Doc PR        | n/a