Maintaining all the RouteCollection abilities to RouteCollectionBuilder
This allows us to add all of these values AFTER all of the routes have been added,
which is the behavior that makes more sense. Also adding a feature so that you can
mount "collections" or routes at a time, and set the controllerClass on those to
save coding routes from an entire class.