Introduction of SyntaxAwareDefinition, SyntaxAwareParameterBag and SyntaxAwareContainerBuilder.
The purpose is to allow for YAML-like @service_name shortcuts through PHP. The
SyntaxAwareDefinition is the most important part of this, but the SyntaxAwareContainerBuilder is also important as it allows you to use the ContainerBuilder::register() fluent interface
for adding these types of Definitions.

The SyntaxAwareParameterBag completes the last use of the shortcuts - expression shortcuts
inside parameters.