Bugfix, issue #16488, Phpunit tests wont run if composer is installed in a wrapper
If composer is installed and found in the path, an interpreter should not be necessary for executing it.
This will help when the composer-executable found in the path is not an actual php-file, but an wrapper for composer.
This is the case when composer have been installed using the homebrew package manager.