[Form] fixed false value being converted to an empty string in ChoiceArrayList
When `false` value is used as a choice value, it is converted to an empty string,
which conflicts with the empty value, if the choice is optional. If the choice
is not optional, then there may be problems with html validation, since required
choice does not allow empty value to be submitted.

This patch disables value casting to string, when the choice contains values,
that would be converted to an empty string.