handleError gets instance of \Error in special cases
For some strange reason i could not discover yet the error_handler is called for catchable fatal errors in PHP 7 with arguments it normally should call the exception_handler.

I have worked around this by declaring the other arguments of handleError as optional an test $type for beeing an instance of Throwable and pass the object to the correct method.

I dont know i you need a example for this case. I have tried to create a small example but i was not able to create a example yet.