[Filesystem] Added 2 new methods, isDir() and isFile().
Those 2 methods are checking if a given path is a directory or a file.

Since we are using Symfony Filesystem component to decouple filesystem from our code, I felt those 2 methods are missing and should be there. People should use them in their code, and then they will be able to mock filesystem dependency in their tests completely.