[Filesystem] Exclude testMirrorCopiesRelativeLinkedContents on Windows
symlink() is not working properly with relative links. There seems to
be some problems with chdir() and symlink() which leads to an error that
symlink cannot read the source. There are some special rare cases
which seems to work if your source points out to the drive root and goes
back into (another) folder, but I would not trust in this case and simply
skip the test.