[DomCrawler] Remove the overridden getHash() method to prevent problems when cloning the crawler
Overriding the SplObjectStorage::getHash() is affected by a PHP bug (https://bugs.php.net/bug.php?id=67582),
which makes the Crawler unusable in Symfony 2.8 for anyone who relied on SplObjectStorage methods.
Removing the getHash() method means we will no longer trigger the deprecation error. Given this method
is unlikely to be used directly and other SplObjectStorage methods will trigger the error, it is the simplest
thing we can do to maintain BC.