[DoctrineBridge][Form] Fix performance regression in EntityType
A performance regression was introduced in
https://github.com/symfony/symfony/commit/2336d5c466325c41b3992d8d3eb0e26d2c4230fa

Before, the default behaviour of the `DoctrineLoader` was to
only fetch the entities selected in the submitted form.

After, the optimization was only performed when the `choice_value`
option was set to `null`.
However, the `DoctrineType` sets a non-null default value to `choice_value`,
which means that the default behaviour was not using the optimization
anymore.

This commit restores the default behaviour (while keeping the previous
commit intent).

References:
- https://github.com/symfony/symfony/blob/v2.7.10/src/Symfony/Bridge/Doctrine/Form/ChoiceList/DoctrineChoiceLoader.php#L149
- https://github.com/symfony/symfony/blob/v2.7.10/src/Symfony/Bridge/Doctrine/Form/Type/DoctrineType.php#L216