[Console] Dispatcher available to extending classes
| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | no
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | yes
| License       | MIT

Currently one can set a event dispatcher on console application but can not access it later.
This makes it impossible for extending classes and/or commands (which have a ref. to the application) to access the dispatcher to register as a listener or to dispatch (custom) events .
This results in application having to keeping a ref. to the dispatcher itself, setup a whole new dispatcher or setup all listeners during construction.