[Process] Allow processes in Unix shells on Windows platform to run in TTY/PTY mode
It is a common use case to use git-bash (or git-shell) as a terminal on Windows platform.
The git-bash, for example, can be run in TTY/PTY mode. Until now, the TTY/PTY mode
cannot be enabled in Unix shells on Windows platform, because the Windows platform
is detected by the 'DIRECTORY_SEPARATOR' constant.

This PR extends the Windows platform detection by also  checking if '/dev/tty' is existing
and readable. If '/dev/tty' is existing and readable and the 'DIRECTORY_SEPARATOR'
constant has Windows platform format, we can assume that the process will be run in
a Unix-like environment (git-shell) on Windows platform and therefore allow TTY/PTY
mode.