[HttpFoundation] fixed testGetContentCanBeCalledTwiceWithResources test
After calling Request::getContent(true), subsequent calls to the
same instance method (withouth the $asResource flag) always returned
false instead of the request body as a plain string.
A unit test already existed to guard against this behaviour but it
yielded a false positive because it was comparing '' to false using
PHPUnit's assertEquals() instead of assertSame().
For completeness sake I also added the missing usage permutations in
the test data provider.