Allow defining a Name Converter by tag
This allows the user to be able to configure a name converter for the
serializer to use. This is useful because it saves you writing a
normalizer for every single class you want to have underscores in the
name of the fields.

With this commit you can tag a service as "serializer.name-converter",
and it will be inserted into the "serializer.normalizer.object" service,
which is the primary service for normalizing general objects into
JSON or whatever. The highest priority name-converter is taken, allowing
the user to override what's been set in there.