[HttpKernel] finding loaded class name when loading container from cache file
When generating multiple versions of the same container class, each new class name is suffixed with an incrementing version number, as you can't re-use class names.
So when we load the class file, the class itself may have a different name than the file name suggests - so after we load the class file, we have to find the class name.