Trim constant values in XmlFileLoader
When an XML config file gets long, it's nice to put the constant name into the next line like this:
```xml
        <parameter key="som_very_very_long_constant_name" type="constant">
            Some\Namespace\That\Is\Sufficiently\Long\To\Require\A\LineBreak::someLongishConstantName
        </parameter>
```

If that’s the case, `constant()` will try and find a constant including the spaces. While it is theoretically possible to have a constant in PHP that starts or ends with a space, it’s impossible for class constants and only doable using `define(" FOO ", …);`. So that’s probably an edge case we can ignore.