[Form] Keep existing entry options when resizing a collection.
Currently you can't edit a collection entry during a form event,
because even if you do so, it will be erased by the ResizeFormListener.

This PR changes this behavior, by reusing options of the previous item
it replaces if it exists.
Because of that I'm not sure which branch to target.
It should be considered a fix because nobody could rely on this "erase" behavior anyway.