[HttpFoundation] Fix Request::preparePathInfo with rewrites
preparePathInfo naively strips strlen($baseUrl) characters off the
front of the requestUri. If the request has passed through url
rewriting by the server, this may strip an incorrect amount of
characters off of the PathInfo causing all routes to mismatch

Additionally, the server may set PATH_INFO which would save us
all of this work in the first place, so use that if we've got it.