refactor setMarking() inside getMarking()
MarkingStore->setMarking() is being called on every call to getMarking() in the Workflow class. 

This is only required if a new Marking is initialised and set as per line 63:
`$marking->mark($this->definition->getInitialPlace());`

If the Marking is not new, the subject either already has a marking and we don't need to call setMarking(), or it is new && theres no `initial place`. In that case an Exception is thrown on line 61.