[Console] Optionally stop options parsing at "--", fixes #21869
Looking for options in the (raw) arguments should not look after an
argument of "--" has been encountered.

Not doing so prevents invoking the command or passing these arguments to
the command w/o interfering with the wrong application or command.

Places where the application looks for global flags like --verbose or
--help make use of the new OPTION_FLAG_POSIX that stops parsing options
at the options-from-operands delimiter "--" as in 12.2 Utility Syntax
Guidelines Guideline 10.

Refs:

- http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap12.html#tag_12_02