Place response date fixup in the forward() method
This makes use of the fact that all requests to the backend go through this method.

Backends should provide a Date header indicating the time at which the response
was generated. But if (for whatever reason) they don't, we'll keep the time
at which the response was received as the response Date.

To-Do: Review the RFC; it probably suggests using the time at which the request
was initiated, not the time the response was received.