[TwigBundle] 'Content-Type' response header match the requested format when generating error pages.
The 'Content-Type' response header need to match the format of the generated response content.

This is obviously useful when an exception is thrown during an XMLHttpRequest handling using JSON content. Currently, the browser (Firefox Developer Edition) handles well the response content and show it as a JSON object. Maybe by looking the request content type or just by recognizing the response content type while parsing it. But the response type is shown as _html_. This PR fix it.

By using the _Request::getMimeType()_ method, we take care of the additional format added by the developer.